Extract all batch execution builtin
