+++ src/builtins/index.ts
@@ -1,3 +1,4 @@
+export * from './all-batch.builtin';
 export * from './command-execution.builtin';
 export * from './parallel-batches';
 export * from './standard-setup';
