+++ src/builtins/standard-setup.ts
@@ -4,12 +4,14 @@
  */
 import { arrayOfElements } from '@proc7ts/primitives';
 import { ZConfig, ZExtension, ZSetup } from '../core';
+import { ZAllBatchBuiltin } from './all-batch.builtin';
 import { ZCommandExecutionBuiltin } from './command-execution.builtin';
 import { ZParallelBatchBuiltin } from './parallel-batches';
 
 const builtinZExtensions: readonly ZExtension[] = [
-    ZParallelBatchBuiltin,
-    ZCommandExecutionBuiltin,
+  ZAllBatchBuiltin,
+  ZCommandExecutionBuiltin,
+  ZParallelBatchBuiltin,
 ];
 
 /**
